Cake v4changes

+++ src/store/actions/user.js
@@ -1,37 +1,45 @@
-import { getRequest, patchRequest, postRequest } from '../../utility/ApiCalls';
+import {
+  getRequest,
+  patchRequest,
+  postRequest,
+} from '../../utility/functions/ApiCalls';
+import { splitUserAndRents } from '../../utility/functions/Helper';
+import { UPDATE_RENT_LIST } from './rents';
 export const USER_LOADING_ERROR = 'USER_LOADING_ERROR';
 export const UPDATE_USER = 'UPDATE_USER';
 
 export const login = (credintials) => {
   return async (dispatch) => {
-    const res = await postRequest(
-      'http://localhost/bookshop/users/login',
-      credintials
-    );
-    if (res?.data?.success) {
-      sessionStorage.setItem('id', res.data.data.id);
-      sessionStorage.setItem('username', res.data.data.username);
-      sessionStorage.setItem('created', res.data.data.created);
-      sessionStorage.setItem('modified', res.data.data.modified);
+    const res = await postRequest('/cake/users/login', credintials);
+    if (res.success && res?.data?.response?.success) {
+      sessionStorage.setItem('id', res.data.response.data.id);
+      sessionStorage.setItem('username', res.data.response.data.username);
+      sessionStorage.setItem('created', res.data.response.data.created);
+      sessionStorage.setItem('modified', res.data.response.data.modified);
+      const { user, rents } = splitUserAndRents(res.data.response.data);
       dispatch({
         type: UPDATE_USER,
-        payload: res.data.data,
+        payload: user,
+      });
+      dispatch({
+        type: UPDATE_RENT_LIST,
+        payload: rents,
       });
       return { success: true };
     } else {
       dispatch({
         type: USER_LOADING_ERROR,
-        payload: res?.data?.message,
+        payload: res?.data?.response?.message,
       });
-      return { success: false, msg: res?.data?.message };
+      return { success: false, msg: res?.data?.response?.message };
     }
   };
 };
 
 export const logout = () => {
   return async (dispatch) => {
-    const res = await getRequest('http://localhost/bookshop/users/logout');
-    if (res?.data?.success) {
+    const res = await getRequest('/cake/users/logout');
+    if (res.success && res?.data?.response?.success) {
       sessionStorage.clear();
       dispatch({
         type: UPDATE_USER,
@@ -40,7 +48,7 @@ export const logout = () => {
     } else {
       dispatch({
         type: USER_LOADING_ERROR,
-        payload: res?.data?.message,
+        payload: res?.data?.response?.message,
       });
     }
   };
@@ -53,58 +61,37 @@ export const setUser = (user) => {
   };
 };
 
-export const getUser = (id) => {
-  return async (dispatch) => {
-    const res = await getRequest(`http://localhost/bookshop/users/view/${id}`);
-    if (res?.data?.success) {
-      dispatch({
-        type: UPDATE_USER,
-        payload: res.data.data,
-      });
-      return { success: true };
-    } else {
-      dispatch({
-        type: USER_LOADING_ERROR,
-        payload: res?.data?.message,
-      });
-      return { success: false, msg: res?.data?.message };
-    }
-  };
-};
-
 export const addUser = (body) => {
   return async (dispatch) => {
-    const res = await postRequest('http://localhost/bookshop/users/add', body);
-    if (res?.data?.success) {
+    const res = await postRequest('/cake/users/add', body);
+    if (res.success && res?.data?.response?.success) {
       return { success: true };
     } else {
       dispatch({
         type: USER_LOADING_ERROR,
-        payload: res?.data?.message,
+        payload: res?.data?.response?.message,
       });
-      return { success: false, msg: res?.data?.message };
+      return { success: false, msg: res?.data?.response?.message };
     }
   };
 };
 
 export const editUser = (id, body) => {
   return async (dispatch) => {
-    const res = await patchRequest(
-      `http://localhost/bookshop/users/edit/${id}`,
-      body
-    );
-    if (res?.success) {
+    const res = await patchRequest(`/cake/users/edit/${id}`, body);
+    if (res?.success && res?.data?.response?.success) {
+      const { user } = splitUserAndRents(res.data.response.data);
       dispatch({
         type: UPDATE_USER,
-        payload: res.data,
+        payload: user,
       });
       return { success: true };
     } else {
       dispatch({
         type: USER_LOADING_ERROR,
-        payload: res?.data?.message,
+        payload: res?.data?.response?.message,
       });
-      return { success: false, msg: res?.data?.message };
+      return { success: false, msg: res?.data?.response?.message };
     }
   };
 };
